Refactor create-project page to handle invalid file types and large files
